Fix browser runtime review findings
