+++ packages/shared-test/black-box-runner/browser/rallar-browser-runtime/runtime.ts
@@ -2,38 +2,27 @@ import type { AuthSession, LoginResponse } from '@shared/api/api-config.ts';
 import type { RtcDataChannelLaneConfig } from '@shared/services/WebRtcConnectionService.ts';
 import {
     createRallarFacade,
-    type RallarCrdtDocument,
-    type RallarCrdtOpenOptions,
-    type RallarDirectorRelayHandle,
-    type RallarDirectorRelayMessage,
-    type RallarDirectorStatus,
     type RallarFacade,
     type RallarRealtimeLaneHealth,
-    type RallarRealtimeSendResult,
     type RallarRtcDiagnostics,
 } from '@shared-web/browser/rallar.ts';
-import { catchUpRallarCrdtDocument } from '@shared-web/browser/api-integration.ts';
-import type {
-    RallarCrdtOperation,
-    RallarCrdtOperationBatch,
-    RallarCrdtSyncOptions,
-    RallarCrdtTransportStrategy,
-    RallarCrdtUpdateEnvelope,
-} from '@shared/crdt/mod.ts';
+import type { RallarCrdtTransportStrategy } from '@shared/crdt/mod.ts';
 import {
     blackBoxRallarAuthenticationIdentityOf,
+    blackBoxRallarConnectionOperationKeyOf,
     blackBoxRallarConnectionTargetOf,
     decideBlackBoxRallarLifecycleRequest,
     isSameBlackBoxRallarSession,
+    mergeBlackBoxRallarAuthenticationConfig,
 } from './policy.ts';
 import {
     createBlackBoxRallarLifecycleController,
     type BlackBoxRallarLifecycleOperationContext,
 } from './lifecycle-controller.ts';
 import { createBlackBoxRallarCrdtController } from './crdt-controller.ts';
-import { createBlackBoxRallarDirectorController, type BlackBoxRallarDirectorLease } from './director-controller.ts';
-import { createBlackBoxRallarMessagingController, type BlackBoxRallarMessagingLease } from './messaging-controller.ts';
-import { createBlackBoxRallarConsoleDiagnostics, createBlackBoxRallarDiagnostics } from './diagnostics.ts';
+import { createBlackBoxRallarDirectorController } from './director-controller.ts';
+import { createBlackBoxRallarMessagingController } from './messaging-controller.ts';
+import { createBlackBoxRallarConsoleDiagnostics, createBlackBoxRallarRuntimeDiagnostics } from './diagnostics.ts';
 
 import type {
     BlackBoxRallarScope,
@@ -46,33 +35,9 @@ import type {
     BlackBoxRallarEvent,
     BlackBoxRallarConnectDiagnostics,
     BlackBoxRallarAuthenticateDiagnostics,
-    BlackBoxRallarSendDiagnostics,
-    BlackBoxRallarWsSendDiagnostics,
     BlackBoxRallarCloseDiagnostics,
     BlackBoxRallarHealthDiagnostics,
     BlackBoxRallarHealthInput,
-    BlackBoxRallarCrdtOpenInput,
-    BlackBoxRallarCrdtHandleInput,
-    BlackBoxRallarCrdtApplyInput,
-    BlackBoxRallarCrdtSyncInput,
-    BlackBoxRallarCrdtWaitOperator,
-    BlackBoxRallarCrdtWaitCondition,
-    BlackBoxRallarCrdtWaitInput,
-    BlackBoxRallarCrdtUndoRedoInput,
-    BlackBoxRallarCrdtRuntimeSummary,
-    BlackBoxRallarCrdtCommandDiagnostics,
-    BlackBoxRallarCrdtRuntime,
-    BlackBoxRallarDirectorRoomInput,
-    BlackBoxRallarDirectorAppointInput,
-    BlackBoxRallarDirectorStatusInput,
-    BlackBoxRallarDirectorRelayStartInput,
-    BlackBoxRallarDirectorHandleInput,
-    BlackBoxRallarDirectorIntentInput,
-    BlackBoxRallarDirectorSyncRequestInput,
-    BlackBoxRallarDirectorOutputRecord,
-    BlackBoxRallarDirectorRelaySummary,
-    BlackBoxRallarDirectorCommandDiagnostics,
-    BlackBoxRallarDirectorRuntime,
     BlackBoxRallarRuntime,
 } from './contracts.ts';
 
@@ -142,17 +107,6 @@ type RuntimeClosePreparation = Readonly<{
     config?: BlackBoxRallarConnectionConfig;
 }>;
 
-type DirectorRelayState = {
-    handle: string;
-    input: BlackBoxRallarDirectorRelayStartInput;
-    relay: RallarDirectorRelayHandle<unknown, BlackBoxRallarDirectorOutputRecord, unknown>;
-    acceptedIntents: unknown[];
-    outputs: unknown[];
-    snapshots: unknown[];
-    syncRequests: unknown[];
-    sequence: number;
-};
-
 declare global {
     interface Window {
         __blackBoxRallar?: BlackBoxRallarRuntime;
@@ -188,6 +142,22 @@ function createBlackBoxRallarRuntimeInstallation(
     let authenticationState: RuntimeAuthenticationState | undefined;
     let authenticationProvenance: RuntimeAuthenticationProvenance | undefined;
     let closeRetryConfig: BlackBoxRallarConnectionConfig | undefined;
+    const runtimeDiagnostics = createBlackBoxRallarRuntimeDiagnostics({
+        now,
+        publish: event => targetWindow.__blackBoxRallarEmit?.(event),
+        onPublishError: error => {
+            console.error('black-box Rallar event sink failed', error);
+        },
+        transportOf,
+        laneIdOf,
+        scopeDiagnostics,
+    });
+    const emit = runtimeDiagnostics.emit;
+    const emitDiagnostic = runtimeDiagnostics.emitDiagnostic;
+    const emitError = runtimeDiagnostics.emitError;
+    const emitConnectPhaseStarted = runtimeDiagnostics.emitConnectPhaseStarted;
+    const emitConnectPhaseCompleted = runtimeDiagnostics.emitConnectPhaseCompleted;
+    const serializeError = runtimeDiagnostics.serializeError;
     const lifecycle = createBlackBoxRallarLifecycleController<
         BlackBoxRallarConnectionConfig,
         LoginResponse | AuthSession,
@@ -195,67 +165,58 @@ function createBlackBoxRallarRuntimeInstallation(
         BlackBoxRallarCloseDiagnostics
     >({
         authenticationKey,
-        mergeAuthenticationConfig: (active, next) => ({
-            ...next,
-            rallar: {
-                ...next.rallar,
-                logoutOnClose: active.rallar.logoutOnClose === true || next.rallar.logoutOnClose === true,
-            },
-        }),
+        mergeAuthenticationConfig: mergeBlackBoxRallarAuthenticationConfig,
         authenticationClosedError,
         connectionClosedError,
     });
-    const crdtController = createBlackBoxRallarCrdtController<RallarCrdtDocument<unknown, RallarCrdtOperationBatch>>({
+    const crdtController = createBlackBoxRallarCrdtController({
         generation: lifecycle.generation,
         isCurrent: lifecycle.isCurrent,
+        facade: rallar,
+        now,
+        delay: wait,
+        currentConnectionConfig: () => state?.config,
+        ensureLiveConnection: ensureCrdtLiveConnection,
+        scopeDiagnostics,
+        emit,
+        emitError,
     });
-    const directorController = createBlackBoxRallarDirectorController<DirectorRelayState>({
+    const directorController = createBlackBoxRallarDirectorController({
         generation: lifecycle.generation,
         isCurrent: lifecycle.isCurrent,
+        facade: rallar,
+        now,
+        requireConfig: () => requireState().config,
+        transportOf,
+        roomRefOf,
+        scopeOf,
+        scopeDiagnostics,
+        emit,
+        emitError,
     });
     const messagingController = createBlackBoxRallarMessagingController({
         generation: lifecycle.generation,
         isCurrent: lifecycle.isCurrent,
-    });
-    const eventDiagnostics = createBlackBoxRallarDiagnostics<BlackBoxRallarEvent>({
-        now,
-        emit: event => {
-            const handler = targetWindow.__blackBoxRallarEmit;
-            if (!handler) {
-                return;
-            }
-            try {
-                void Promise.resolve(handler(event)).catch(error => {
-                    console.error('black-box Rallar event sink failed', error);
-                });
-            } catch (error) {
-                console.error('black-box Rallar event sink failed', error);
-            }
-        },
+        facade: rallar,
+        requireConfig: () => requireState().config,
+        transportOf,
+        laneIdOf,
+        typeIdOf,
+        topicIdOf,
+        roomRefOf,
+        scopeDiagnostics,
+        toOptionalNumber,
+        readHealth,
+        wsStatus: wsStatusFor,
+        rtcStatus: rtcStatusFor,
+        emit,
+        emitDiagnostic,
+        emitError,
     });
     const consoleDiagnostics = createBlackBoxRallarConsoleDiagnostics<BlackBoxRallarConnectionConfig>({
         console,
         activeConfig: () => state?.config,
-        onWarning: (activeConfig, args) => {
-            const classified = classifyConsoleWarning(args);
-            if (!classified) {
-                return;
-            }
-            emit({
-                kind: 'diagnostic',
-                topic: classified.topic,
-                connection: activeConfig.connection,
-                actor: activeConfig.actor,
-                transport: classified.transport,
-                severity: 'warning',
-                roomId: activeConfig.roomId,
-                ...scopeDiagnostics(activeConfig),
-                data: {
-                    message: classified.message,
-                    args,
-                },
-            });
-        },
+        onWarning: runtimeDiagnostics.emitConsoleWarning,
         restoreExisting: restoreExistingConsoleWarnPatch,
         publishRestore: restore => {
             consoleWarnGlobalState().__blackBoxRallarRestoreConsoleWarn = restore;
@@ -285,32 +246,6 @@ function createBlackBoxRallarRuntimeInstallation(
 
     function isRecord(value: unknown): value is Record<string, unknown> {
         return Boolean(value) && typeof value === 'object' && !Array.isArray(value);
-    }
-
-    function asRecord(value: unknown): Record<string, unknown> {
-        return isRecord(value) ? value : {};
-    }
-
-    function optionalRecord(value: unknown): Record<string, unknown> | undefined {
-        return isRecord(value) ? value : undefined;
-    }
-
-    function stringValue(value: unknown): string | undefined {
-        return typeof value === 'string' && value.trim().length > 0 ? value : undefined;
-    }
-
-    function wsScopeValue(value: unknown): 'room' | 'world' | 'all' | undefined {
-        return value === 'room' || value === 'world' || value === 'all' ? value : undefined;
-    }
-
-    function maybeStringArray(value: unknown): readonly string[] | undefined {
-        return Array.isArray(value)
-            ? value.filter((entry): entry is string => typeof entry === 'string' && entry.length > 0)
-            : undefined;
-    }
-
-    function wsSelectorKey(typeId: string, topicId?: string): string {
-        return `${topicId ?? '*'}/${typeId}`;
     }
 
     function scopeOf(
@@ -437,777 +372,6 @@ function createBlackBoxRallarRuntimeInstallation(
         return defaults;
     }
 
-    function crdtHandle(input: unknown): string {
-        const record = asRecord(input);
-        const handle = stringValue(record.handle) ?? stringValue(record.commandId) ?? stringValue(record.name);
-        if (!handle) {
-            throw new Error('CRDT command requires handle.');
-        }
-
-        return handle;
-    }
-
-    function crdtRoomRef(input: BlackBoxRallarCrdtOpenInput): BlackBoxRallarRoomRef | undefined {
-        const explicit = input.roomRef ?? (optionalRecord(input.rallar?.roomRef) as BlackBoxRallarRoomRef | undefined);
-        if (explicit?.applicationId && explicit.groupId) {
-            return {
-                applicationId: String(explicit.applicationId),
-                ...(explicit.workspaceId !== undefined ? { workspaceId: String(explicit.workspaceId) } : {}),
-                groupId: String(explicit.groupId),
-            };
-        }
-
-        const roomId = input.roomId ?? stringValue(input.rallar?.roomId);
-        const applicationId =
-            input.applicationId ??
-            stringValue(input.rallar?.applicationId) ??
-            stringValue(optionalRecord(input.scope)?.applicationId);
-        if (!roomId || !applicationId) {
-            return undefined;
-        }
-
-        const workspaceId =
-            input.workspaceId ??
-            stringValue(input.rallar?.workspaceId) ??
-            stringValue(optionalRecord(input.scope)?.workspaceId);
-
-        return {
-            applicationId,
-            ...(workspaceId !== undefined ? { workspaceId } : {}),
-            groupId: roomId,
-        };
-    }
-
-    function normalizeCrdtOpenInput(input: BlackBoxRallarCrdtOpenInput | unknown): BlackBoxRallarCrdtOpenInput {
-        const record = asRecord(input);
-        const rallarConfig = asRecord(record.rallar);
-        const scope = optionalRecord(record.scope) ?? optionalRecord(rallarConfig.scope);
-        const roomRef = optionalRecord(record.roomRef) ?? optionalRecord(rallarConfig.roomRef);
-        const name = stringValue(record.name);
-        if (!name) {
-            throw new Error('crdt.open requires name.');
-        }
-
-        return {
-            handle: stringValue(record.handle) ?? stringValue(record.commandId) ?? name,
-            name,
-            applicationId: stringValue(record.applicationId) ?? stringValue(rallarConfig.applicationId),
-            workspaceId: stringValue(record.workspaceId) ?? stringValue(rallarConfig.workspaceId),
-            documentId: stringValue(record.documentId),
-            documentType: stringValue(record.documentType),
-            scope,
-            roomRef: roomRef as BlackBoxRallarRoomRef | undefined,
-            principalId: stringValue(record.principalId),
-            customScope: stringValue(record.customScope),
-            transport: toCrdtTransport(record.transport) ?? toCrdtTransport(rallarConfig.crdtTransport),
-            persist: typeof record.persist === 'boolean' ? record.persist : undefined,
-            tabSync: typeof record.tabSync === 'boolean' ? record.tabSync : undefined,
-            initialValue: record.initialValue,
-            policies: Array.isArray(record.policies)
-                ? (record.policies as readonly Readonly<Record<string, unknown>>[])
-                : undefined,
-            validation: optionalRecord(record.validation),
-            encryption: optionalRecord(record.encryption),
-            durableCatchUp:
-                record.durableCatchUp === 'http' ? 'http' : record.durableCatchUp === false ? false : undefined,
-            apiBaseUrl: stringValue(record.apiBaseUrl) ?? stringValue(rallarConfig.apiBaseUrl),
-            actor: stringValue(record.actor),
-            sessionId: stringValue(record.sessionId) ?? stringValue(rallarConfig.sessionId),
-            username: stringValue(record.username) ?? stringValue(rallarConfig.username),
-            password: stringValue(record.password) ?? stringValue(rallarConfig.password),
-            displayName: stringValue(record.displayName) ?? stringValue(rallarConfig.displayName),
-            register:
-                record.register === true || record.register === 'if-needed'
-                    ? record.register
-                    : rallarConfig.register === true || rallarConfig.register === 'if-needed'
-                      ? rallarConfig.register
-                      : undefined,
-            timeoutMs:
-                typeof record.timeoutMs === 'number'
-                    ? record.timeoutMs
-                    : typeof rallarConfig.timeoutMs === 'number'
-                      ? rallarConfig.timeoutMs
-                      : undefined,
-            roomId: stringValue(record.roomId) ?? stringValue(rallarConfig.roomId),
-            rallar: rallarConfig,
-        };
-    }
-
-    function toCrdtTransport(value: unknown): RallarCrdtTransportStrategy | undefined {
-        return value === 'local-only' ||
-            value === 'ws' ||
-            value === 'rtc' ||
-            value === 'ws-then-rtc' ||
-            value === 'rtc-with-ws-fallback'
-            ? value
-            : undefined;
-    }
-
-    function toCrdtOpenScope(input: BlackBoxRallarCrdtOpenInput): RallarCrdtOpenOptions['scope'] {
-        const scope = optionalRecord(input.scope);
-        const kind = stringValue(scope?.kind);
-        const roomRef = crdtRoomRef(input);
-        if (kind === 'app') {
-            return { kind: 'app' };
-        }
-        if (kind === 'principal') {
-            const principalId = stringValue(scope?.principalId) ?? input.principalId;
-            if (principalId) {
-                return { kind: 'principal', principalId };
-            }
-        }
-        if (kind === 'custom') {
-            const customScope = stringValue(scope?.customScope) ?? input.customScope;
-            if (customScope) {
-                return { kind: 'custom', customScope };
-            }
-        }
-        if (kind === 'room' && roomRef) {
-            return { kind: 'room', roomRef };
-        }
-        if (input.principalId) {
-            return { kind: 'principal', principalId: input.principalId };
-        }
-        if (input.customScope) {
-            return { kind: 'custom', customScope: input.customScope };
-        }
-        if (roomRef) {
-            return { kind: 'room', roomRef };
-        }
-
-        return undefined;
-    }
-
-    function toCrdtConnectionConfig(input: BlackBoxRallarCrdtOpenInput): BlackBoxRallarConnectionConfig {
-        const roomRef = crdtRoomRef(input);
-        const applicationId = input.applicationId ?? stringValue(input.rallar?.applicationId) ?? roomRef?.applicationId;
-        const workspaceId =
-            input.workspaceId ?? stringValue(input.rallar?.workspaceId) ?? roomRef?.workspaceId ?? DEFAULT_WORKSPACE_ID;
-        const scope = applicationId
-            ? {
-                  applicationId,
-                  workspaceId,
-              }
-            : undefined;
-
-        return {
-            connection: input.handle ?? input.name,
-            actor: input.actor,
-            roomId: input.roomId ?? roomRef?.groupId,
-            ...(roomRef ? { roomRef } : {}),
-            rallar: {
-                ...asRecord(input.rallar),
-                apiBaseUrl: input.apiBaseUrl ?? stringValue(input.rallar?.apiBaseUrl) ?? '',
-                ...(input.username ? { username: input.username } : {}),
-                ...(input.password ? { password: input.password } : {}),
-                ...(input.displayName ? { displayName: input.displayName } : {}),
-                ...(input.register !== undefined ? { register: input.register } : {}),
-                ...(applicationId ? { applicationId } : {}),
-                ...(workspaceId ? { workspaceId } : {}),
-                ...(scope ? { scope } : {}),
-                ...(roomRef ? { roomRef } : {}),
-                ...(input.sessionId ? { expectedSessionId: input.sessionId } : {}),
-                ...(input.timeoutMs !== undefined ? { timeoutMs: input.timeoutMs } : {}),
-                transport: 'realtime',
-            },
-        };
-    }
-
-    function toCrdtOpenOptions(
-        input: BlackBoxRallarCrdtOpenInput,
-    ): RallarCrdtOpenOptions<unknown, RallarCrdtOperationBatch> {
-        return {
-            ...(input.applicationId ? { applicationId: input.applicationId } : {}),
-            ...(input.workspaceId ? { workspaceId: input.workspaceId } : {}),
-            ...(input.documentId ? { documentId: input.documentId } : {}),
-            ...(input.documentType ? { documentType: input.documentType } : {}),
-            ...(toCrdtOpenScope(input) ? { scope: toCrdtOpenScope(input) } : {}),
-            ...(input.transport ? { transport: input.transport } : {}),
-            ...(input.persist !== undefined ? { persist: input.persist } : {}),
-            ...(input.tabSync !== undefined ? { tabSync: input.tabSync } : {}),
-            ...(input.initialValue !== undefined ? { initialValue: input.initialValue } : {}),
-            ...(input.policies ? { policies: input.policies as any } : {}),
-            ...(input.validation ? { validation: input.validation as any } : {}),
-            ...(input.encryption ? { encryption: input.encryption as any } : {}),
-            ...(input.actor ? { actorId: input.actor } : {}),
-            ...(input.sessionId ? { sessionId: input.sessionId } : {}),
-            ...(input.durableCatchUp === 'http' ? { durableCatchUp: catchUpRallarCrdtDocument } : {}),
-        };
-    }
-
-    async function ensureCrdtLiveConnection(
-        input: BlackBoxRallarCrdtOpenInput,
-    ): Promise<BlackBoxRallarConnectionConfig | undefined> {
-        if ((input.transport ?? 'local-only') === 'local-only') {
-            return undefined;
-        }
-        if (!input.apiBaseUrl) {
-            if (rallar.isConnected()) {
-                return state?.config;
-            }
-            throw new Error('crdt.open requires apiBaseUrl or an existing Rallar connection for live transports.');
-        }
-
-        const config = toCrdtConnectionConfig(input);
-        if (lifecycle.authenticationConfig()) {
-            await lifecycle.waitForAuthentication();
-            return await ensureCrdtLiveConnection(input);
-        }
-        if (state) {
-            const decision = decideBlackBoxRallarLifecycleRequest(
-                {
-                    status: 'connected',
-                    activeTarget: blackBoxRallarConnectionTargetOf(state.config, state.session),
-                },
-                {
-                    kind: 'connect',
-                    target: blackBoxRallarConnectionTargetOf(config, state.session),
-                },
-            );
-            if (decision.kind === 'reject') {
-                throw new Error(decision.reason);
-            }
-        }
-        return await lifecycle.runExclusive('crdt-live:' + connectionKey(config), async context => {
-            const queuedRejection = connectedTargetRejection(config);
-            if (queuedRejection) {
-                throw queuedRejection;
-            }
-            emitDiagnostic(config, 'rallar.browser.crdt.configure_started', {
-                transportStrategy: input.transport,
-            });
-            const defaults = configureRallarConnection(config);
-            emitDiagnostic(config, 'rallar.browser.crdt.configure_completed', {
-                defaults,
-            });
-
-            if (!rallar.isConnected()) {
-                const session = await sessionForAuthentication(config);
-                context.assertCurrent();
-                await rallar.connect({
-                    timeoutMs: input.timeoutMs,
-                });
-                context.assertCurrent();
-                if (config.roomId) {
-                    await rallar.rooms.join(config.roomId, {
-                        timeoutMs: input.timeoutMs,
-                        scope: scopeOf(config),
-                    });
-                    context.assertCurrent();
-                }
-                emitDiagnostic(config, 'rallar.browser.crdt.connected', {
-                    session: toSessionDiagnostic(session),
-                    transportStrategy: input.transport,
-                });
-            }
-
-            return config;
-        });
-    }
-
-    function requireCrdtDocument(handle: string): RallarCrdtDocument<unknown, RallarCrdtOperationBatch> {
-        return crdtController.require(handle);
-    }
-
-    function optionalNumber(value: unknown): number | undefined {
-        return typeof value === 'number' && Number.isFinite(value) ? value : undefined;
-    }
-
-    function normalizeCrdtWaitCondition(value: unknown): BlackBoxRallarCrdtWaitCondition {
-        const record = asRecord(value);
-        const source = record.source === 'value' || record.source === 'health' ? record.source : undefined;
-        const operator = isCrdtWaitOperator(record.operator) ? record.operator : undefined;
-        if (!source || !operator) {
-            throw new Error('crdt.wait conditions require source and supported operator.');
-        }
-
-        return {
-            source,
-            ...(stringValue(record.path) ? { path: stringValue(record.path) } : {}),
-            operator,
-            ...(record.expected !== undefined ? { expected: record.expected } : {}),
-        };
-    }
-
-    function normalizeCrdtWaitInput(input: BlackBoxRallarCrdtWaitInput | unknown): BlackBoxRallarCrdtWaitInput {
-        const record = asRecord(input);
-        const handle = crdtHandle(record);
-        const conditions = Array.isArray(record.conditions) ? record.conditions.map(normalizeCrdtWaitCondition) : [];
-        if (conditions.length === 0) {
-            throw new Error('crdt.wait requires at least one condition.');
-        }
-
-        const syncRecord = record.sync === false ? false : optionalRecord(record.sync);
-
-        return {
-            handle,
-            ...(optionalNumber(record.timeoutMs) !== undefined
-                ? {
-                      timeoutMs: Math.max(0, optionalNumber(record.timeoutMs) as number),
-                  }
-                : {}),
-            ...(optionalNumber(record.intervalMs) !== undefined
-                ? {
-                      intervalMs: Math.max(0, optionalNumber(record.intervalMs) as number),
-                  }
-                : {}),
-            ...(optionalNumber(record.stableForMs) !== undefined
-                ? {
-                      stableForMs: Math.max(0, optionalNumber(record.stableForMs) as number),
-                  }
-                : {}),
-            ...(syncRecord !== undefined
-                ? {
-                      sync:
-                          syncRecord === false
-                              ? false
-                              : {
-                                    ...(stringValue(syncRecord.reason)
-                                        ? {
-                                              reason: stringValue(syncRecord.reason),
-                                          }
-                                        : {}),
-                                    ...(toCrdtTransport(syncRecord.transport)
-                                        ? {
-                                              transport: toCrdtTransport(syncRecord.transport),
-                                          }
-                                        : {}),
-                                },
-                  }
-                : {}),
-            conditions,
-        };
-    }
-
-    function isCrdtWaitOperator(value: unknown): value is BlackBoxRallarCrdtWaitOperator {
-        return (
-            value === 'equals' ||
-            value === 'notEquals' ||
-            value === 'contains' ||
-            value === 'exists' ||
-            value === 'gte' ||
-            value === 'lte'
-        );
-    }
-
-    function normalizeCrdtWaitPath(path: string): string {
-        if (path.startsWith('$.')) {
-            return path.slice('$.'.length);
-        }
-        return path;
-    }
-
-    function lookupCrdtWaitPath(
-        root: unknown,
-        path: string | undefined,
-    ): Readonly<{
-        exists: boolean;
-        value?: unknown;
-    }> {
-        if (!path || path.trim().length === 0) {
-            return {
-                exists: root !== undefined,
-                value: root,
-            };
-        }
-
-        let current = root;
-        const segments = normalizeCrdtWaitPath(path)
-            .split('.')
-            .filter(segment => segment.length > 0);
-        for (const segment of segments) {
-            if ((Array.isArray(current) || typeof current === 'string') && segment === 'length') {
-                current = current.length;
-                continue;
-            }
-
-            if (Array.isArray(current)) {
-                const index = Number(segment);
-                if (!Number.isInteger(index) || index < 0 || index >= current.length) {
-                    return { exists: false };
-                }
-                current = current[index];
-                continue;
-            }
-
-            if (!current || typeof current !== 'object') {
-                return { exists: false };
-            }
-
-            const record = current as Record<string, unknown>;
-            if (!Object.prototype.hasOwnProperty.call(record, segment)) {
-                return { exists: false };
-            }
-            current = record[segment];
-        }
-
-        return {
-            exists: true,
-            value: current,
-        };
-    }
-
-    function sameCrdtWaitValue(left: unknown, right: unknown): boolean {
-        try {
-            return JSON.stringify(left) === JSON.stringify(right);
-        } catch (_error) {
-            return Object.is(left, right);
-        }
-    }
-
-    function containsCrdtWaitValue(value: unknown, expected: unknown): boolean {
-        if (Array.isArray(value)) {
-            return value.some(entry => sameCrdtWaitValue(entry, expected));
-        }
-        if (typeof value === 'string') {
-            return value.includes(String(expected));
-        }
-        if (value && typeof value === 'object') {
-            if (typeof expected === 'string') {
-                try {
-                    return JSON.stringify(value).includes(expected);
-                } catch (_error) {
-                    return String(value).includes(expected);
-                }
-            }
-            return Object.values(value as Record<string, unknown>).some(entry => sameCrdtWaitValue(entry, expected));
-        }
-
-        return String(value).includes(String(expected));
-    }
-
-    function crdtWaitConditionMatches(
-        condition: BlackBoxRallarCrdtWaitCondition,
-        value: unknown,
-        health: unknown,
-    ): boolean {
-        const source = condition.source === 'value' ? value : health;
-        const lookup = lookupCrdtWaitPath(source, condition.path);
-
-        switch (condition.operator) {
-            case 'equals':
-                return lookup.exists && sameCrdtWaitValue(lookup.value, condition.expected);
-            case 'notEquals':
-                return !lookup.exists || !sameCrdtWaitValue(lookup.value, condition.expected);
-            case 'contains':
-                return lookup.exists && containsCrdtWaitValue(lookup.value, condition.expected);
-            case 'exists':
-                return condition.expected === undefined ? lookup.exists : lookup.exists === Boolean(condition.expected);
-            case 'gte':
-                return (
-                    lookup.exists &&
-                    typeof lookup.value === 'number' &&
-                    typeof condition.expected === 'number' &&
-                    lookup.value >= condition.expected
-                );
-            case 'lte':
-                return (
-                    lookup.exists &&
-                    typeof lookup.value === 'number' &&
-                    typeof condition.expected === 'number' &&
-                    lookup.value <= condition.expected
-                );
-        }
-    }
-
-    function crdtRuntimeSummary(): BlackBoxRallarCrdtRuntimeSummary {
-        return {
-            handles: crdtController.handles(),
-            documents: crdtController.entries().map(([handle, document]) => ({
-                handle,
-                ref: document.ref,
-                health: document.health(),
-            })),
-        };
-    }
-
-    function normalizeDirectorRoomInput(input: unknown): BlackBoxRallarDirectorRoomInput {
-        const record = asRecord(input);
-        const rallarConfig = asRecord(record.rallar);
-        const scope = optionalRecord(record.scope) ?? optionalRecord(rallarConfig.scope);
-        const roomRef = optionalRecord(record.roomRef) ?? optionalRecord(rallarConfig.roomRef);
-        return {
-            roomId: stringValue(record.roomId) ?? stringValue(record.groupId) ?? stringValue(rallarConfig.roomId),
-            applicationId: stringValue(record.applicationId) ?? stringValue(rallarConfig.applicationId),
-            workspaceId: stringValue(record.workspaceId) ?? stringValue(rallarConfig.workspaceId),
-            scope,
-            roomRef: roomRef as BlackBoxRallarRoomRef | undefined,
-            timeoutMs: optionalNumber(record.timeoutMs),
-        };
-    }
-
-    function normalizeDirectorAppointInput(input: unknown): BlackBoxRallarDirectorAppointInput {
-        const record = asRecord(input);
-        return {
-            ...normalizeDirectorRoomInput(record),
-            heartbeatTtlMs: optionalNumber(record.heartbeatTtlMs),
-        };
-    }
-
-    function normalizeDirectorStatusInput(input: unknown): BlackBoxRallarDirectorStatusInput {
-        const record = asRecord(input);
-        return {
-            ...normalizeDirectorRoomInput(record),
-            refresh: typeof record.refresh === 'boolean' ? record.refresh : undefined,
-            now: optionalNumber(record.now),
-        };
-    }
-
-    function normalizeDirectorRelayStartInput(input: unknown): BlackBoxRallarDirectorRelayStartInput {
-        const record = asRecord(input);
-        const handle = stringValue(record.handle);
-        const intentTypeId = stringValue(record.intentTypeId);
-        const outputTypeId = stringValue(record.outputTypeId);
-        if (!handle || !intentTypeId || !outputTypeId) {
-            throw new Error('director.relay.start requires handle, intentTypeId, and outputTypeId.');
-        }
-
-        return {
-            ...normalizeDirectorRoomInput(record),
-            handle,
-            laneId: stringValue(record.laneId),
-            topicId: stringValue(record.topicId),
-            intentTypeId,
-            outputTypeId,
-            heartbeatTypeId: stringValue(record.heartbeatTypeId),
-            snapshotTypeId: stringValue(record.snapshotTypeId),
-            syncRequestTypeId: stringValue(record.syncRequestTypeId),
-            heartbeatIntervalMs: optionalNumber(record.heartbeatIntervalMs),
-            snapshotIntervalMs: optionalNumber(record.snapshotIntervalMs),
-            ...(Object.prototype.hasOwnProperty.call(record, 'snapshot') ? { snapshot: record.snapshot } : {}),
-        };
-    }
-
-    function normalizeDirectorHandleInput(input: unknown): BlackBoxRallarDirectorHandleInput {
-        const record = asRecord(input);
-        const handle = stringValue(record.handle);
-        if (!handle) {
-            throw new Error('Director relay command requires handle.');
-        }
-        return {
-            handle,
-            timeoutMs: optionalNumber(record.timeoutMs),
-        };
-    }
-
-    function normalizeDirectorIntentInput(input: unknown): BlackBoxRallarDirectorIntentInput {
-        const record = asRecord(input);
-        return {
-            ...normalizeDirectorHandleInput(record),
-            intent: record.intent,
-        };
-    }
-
-    function normalizeDirectorSyncRequestInput(input: unknown): BlackBoxRallarDirectorSyncRequestInput {
-        const record = asRecord(input);
-        return {
-            ...normalizeDirectorHandleInput(record),
-            payload: record.payload,
-        };
-    }
-
-    function toDirectorRoomRef(
-        input: BlackBoxRallarDirectorRoomInput,
-        config: BlackBoxRallarConnectionConfig,
-    ): BlackBoxRallarRoomRef | undefined {
-        return input.roomRef ?? roomRefOf(config, input as BlackBoxRallarSendInput);
-    }
-
-    function toDirectorTarget(
-        input: BlackBoxRallarDirectorRoomInput,
-        config: BlackBoxRallarConnectionConfig,
-    ): string | BlackBoxRallarRoomRef | undefined {
-        return toDirectorRoomRef(input, config) ?? input.roomId ?? config.roomId;
-    }
-
-    function toDirectorScope(
-        input: BlackBoxRallarDirectorRoomInput,
-        config: BlackBoxRallarConnectionConfig,
-    ): ResolvedBlackBoxRallarScope | undefined {
-        return scopeOf(config, input as BlackBoxRallarSendInput);
-    }
-
-    function directorStatusDiagnostics(
-        status: BlackBoxRallarDirectorCommandDiagnostics['status'],
-        input: BlackBoxRallarDirectorRoomInput,
-        directorStatus: RallarDirectorStatus,
-        config: BlackBoxRallarConnectionConfig,
-        extra: Omit<
-            BlackBoxRallarDirectorCommandDiagnostics,
-            | 'status'
-            | 'roomId'
-            | 'roomRef'
-            | 'role'
-            | 'state'
-            | 'isDirector'
-            | 'isFresh'
-            | 'appointment'
-            | 'directorStatus'
-        > = {},
-    ): BlackBoxRallarDirectorCommandDiagnostics {
-        const roomRef = toDirectorRoomRef(input, config);
-        return {
-            status,
-            roomId: input.roomId ?? roomRef?.groupId ?? config.roomId ?? directorStatus.roomId,
-            ...(roomRef ? { roomRef } : {}),
-            role: directorStatus.role,
-            state: directorStatus.state,
-            isDirector: directorStatus.isDirector,
-            isFresh: directorStatus.isFresh,
-            appointment: directorStatus.appointment,
-            directorStatus,
-            ...extra,
-        };
-    }
-
-    function directorRelaySnapshot(entry: DirectorRelayState): Record<string, unknown> {
-        const status = entry.relay.status();
-        return entry.input.snapshot !== undefined
-            ? {
-                  handle: entry.handle,
-                  static: true,
-                  status,
-                  snapshot: entry.input.snapshot,
-              }
-            : {
-                  handle: entry.handle,
-                  status,
-                  acceptedIntents: entry.acceptedIntents,
-                  outputs: entry.outputs,
-                  snapshots: entry.snapshots,
-                  syncRequests: entry.syncRequests,
-                  sequence: entry.sequence,
-                  generatedAtEpochMs: now(),
-              };
-    }
-
-    function directorRelaySummary(): BlackBoxRallarDirectorRelaySummary {
-        return {
-            handles: directorController.handles(),
-            relays: directorController.entries().map(([, entry]) => ({
-                handle: entry.handle,
-                roomId: entry.input.roomId,
-                topicId: entry.input.topicId,
-                intentTypeId: entry.input.intentTypeId,
-                outputTypeId: entry.input.outputTypeId,
-                acceptedIntentCount: entry.acceptedIntents.length,
-                outputCount: entry.outputs.length,
-                snapshotCount: entry.snapshots.length,
-                syncRequestCount: entry.syncRequests.length,
-                status: entry.relay.status(),
-            })),
-        };
-    }
-
-    function emitDirectorDiagnostic(
-        topic: string,
-        handle: string | undefined,
-        data: unknown,
-        config: BlackBoxRallarConnectionConfig,
-    ): void {
-        emit({
-            kind: 'diagnostic',
-            topic,
-            connection: config.connection,
-            actor: config.actor,
-            transport: transportOf(config),
-            roomId: config.roomId,
-            ...scopeDiagnostics(config),
-            data: {
-                ...(handle ? { handle } : {}),
-                ...(data && typeof data === 'object' && !Array.isArray(data)
-                    ? (data as Record<string, unknown>)
-                    : { value: data }),
-            },
-        });
-    }
-
-    function intentIdFromPayload(payload: unknown, fallback: string): string {
-        const record = asRecord(payload);
-        return stringValue(record.intentId) ?? stringValue(record.id) ?? stringValue(record.messageId) ?? fallback;
-    }
-
-    function requireDirectorRelay(handle: string): DirectorRelayState {
-        try {
-            return directorController.require(handle);
-        } catch {
-            throw new Error('Director relay handle is not open: ' + handle);
-        }
-    }
-
-    function assertDirectorCurrent(lease: BlackBoxRallarDirectorLease): void {
-        directorController.assertCurrent(lease, 'Director operation completed after the runtime closed.');
-    }
-
-    function toCrdtDiagnostics(
-        status: BlackBoxRallarCrdtCommandDiagnostics['status'],
-        handle: string,
-        document: RallarCrdtDocument<unknown, RallarCrdtOperationBatch> | undefined,
-        options: Readonly<{
-            update?: RallarCrdtUpdateEnvelope<RallarCrdtOperationBatch>;
-            result?: unknown;
-            value?: unknown;
-            transportStrategy?: RallarCrdtTransportStrategy;
-            attempts?: number;
-            waitedMs?: number;
-            stableForMs?: number;
-            conditions?: readonly BlackBoxRallarCrdtWaitCondition[];
-            lastSyncResult?: unknown;
-        }> = {},
-    ): BlackBoxRallarCrdtCommandDiagnostics {
-        const health = document?.health();
-        const value =
-            options.value !== undefined
-                ? options.value
-                : document && status !== 'closed' && status !== 'destroyed'
-                  ? document.read()
-                  : undefined;
-
-        return {
-            status,
-            handle,
-            ...(document ? { ref: document.ref } : {}),
-            ...(options.transportStrategy ? { transportStrategy: options.transportStrategy } : {}),
-            ...(options.update ? { updateId: options.update.updateId } : {}),
-            ...(value !== undefined ? { value } : {}),
-            ...(options.result !== undefined ? { result: options.result } : {}),
-            ...(health !== undefined ? { health } : {}),
-            ...(document ? { pendingUpdateCount: document.pendingUpdates().length } : {}),
-            ...(document
-                ? {
-                      failedPendingUpdateCount: document.failedPendingUpdates().length,
-                  }
-                : {}),
-            ...(document
-                ? {
-                      dependencyBlockedUpdateCount: document.dependencyBlockedUpdates().length,
-                  }
-                : {}),
-            ...(options.attempts !== undefined ? { attempts: options.attempts } : {}),
-            ...(options.waitedMs !== undefined ? { waitedMs: options.waitedMs } : {}),
-            ...(options.stableForMs !== undefined ? { stableForMs: options.stableForMs } : {}),
-            ...(options.conditions ? { conditions: options.conditions } : {}),
-            ...(options.lastSyncResult !== undefined ? { lastSyncResult: options.lastSyncResult } : {}),
-        };
-    }
-
-    function emitCrdtDiagnostic(
-        topic: string,
-        handle: string,
-        data: unknown,
-        config?: BlackBoxRallarConnectionConfig,
-    ): void {
-        emit({
-            kind: 'diagnostic',
-            topic,
-            connection: config?.connection ?? handle,
-            actor: config?.actor,
-            roomId: config?.roomId,
-            ...(config ? scopeDiagnostics(config) : {}),
-            data,
-        });
-    }
-
     function messageSelectorOf(config: BlackBoxRallarConnectionConfig):
         | string
         | {
@@ -1224,24 +388,6 @@ function createBlackBoxRallarRuntimeInstallation(
         };
     }
 
-    function emit(partial: Omit<BlackBoxRallarEvent, 'atEpochMs'>): void {
-        eventDiagnostics.emit(partial);
-    }
-
-    function emitDiagnostic(config: BlackBoxRallarConnectionConfig, topic: string, data?: unknown): void {
-        emit({
-            kind: 'diagnostic',
-            topic,
-            connection: config.connection,
-            actor: config.actor,
-            transport: transportOf(config),
-            roomId: config.roomId,
-            ...scopeDiagnostics(config),
-            laneId: laneIdOf(config),
-            data,
-        });
-    }
-
     function installConsoleDiagnostics(config: BlackBoxRallarConnectionConfig): () => void {
         return consoleDiagnostics.install(config);
     }
@@ -1256,79 +402,6 @@ function createBlackBoxRallarRuntimeInstallation(
         return globalThis as typeof globalThis & {
             __blackBoxRallarRestoreConsoleWarn?: () => void;
         };
-    }
-
-    function classifyConsoleWarning(args: readonly unknown[]):
-        | Readonly<{
-              topic: string;
-              transport: BlackBoxRallarTransport | 'ws';
-              message: string;
-          }>
-        | undefined {
-        const message = args.map(consoleWarningPart).join(' ');
-        if (message.includes('Unhandled WS message') || message.includes('No callback for typeId')) {
-            return {
-                topic: 'rallar.browser.ws.unhandled_message',
-                transport: 'ws',
-                message,
-            };
-        }
-        if (
-            message.includes('Received data channel for different data channel name') ||
-            message.includes('does not match peerId') ||
-            message.includes('No channel for peer') ||
-            message.includes('Ignoring self-connection attempt')
-        ) {
-            return {
-                topic: 'rallar.browser.rtc.data_channel_warning',
-                transport: 'realtime',
-                message,
-            };
-        }
-        return undefined;
-    }
-
-    function consoleWarningPart(value: unknown): string {
-        if (typeof value === 'string') {
-            return value;
-        }
-        try {
-            return JSON.stringify(value);
-        } catch (_error) {
-            return String(value);
-        }
-    }
-
-    function emitError(
-        config: BlackBoxRallarConnectionConfig | undefined,
-        topic: string,
-        error: unknown,
-        data?: unknown,
-    ): void {
-        emit({
-            kind: 'diagnostic',
-            topic,
-            connection: config?.connection,
-            actor: config?.actor,
-            transport: config ? transportOf(config) : undefined,
-            roomId: config?.roomId,
-            ...(config ? scopeDiagnostics(config) : {}),
-            laneId: config ? laneIdOf(config) : undefined,
-            data,
-            error: serializeError(error),
-        });
-    }
-
-    function serializeError(error: unknown): unknown {
-        if (error instanceof Error) {
-            return {
-                name: error.name,
-                message: error.message,
-                stack: error.stack,
-            };
-        }
-
-        return error;
     }
 
     function toSessionDiagnostic(session: RuntimeSessionDiagnostic): RuntimeSessionDiagnostic {
@@ -1407,10 +480,6 @@ function createBlackBoxRallarRuntimeInstallation(
         });
     }
 
-    function toDiagnosticObject(data?: unknown): Record<string, unknown> {
-        return data && typeof data === 'object' && !Array.isArray(data) ? (data as Record<string, unknown>) : { data };
-    }
-
     function toOptionalNumber(value: unknown): number | undefined {
         if (value === undefined || value === null || value === '') {
             return undefined;
@@ -1418,20 +487,6 @@ function createBlackBoxRallarRuntimeInstallation(
 
         const parsed = typeof value === 'number' ? value : Number(value);
         return Number.isFinite(parsed) ? parsed : undefined;
-    }
-
-    function emitConnectPhaseStarted(config: BlackBoxRallarConnectionConfig, phase: string, data?: unknown): void {
-        emitDiagnostic(config, 'rallar.browser.connect.phase_started', {
-            phase,
-            ...toDiagnosticObject(data),
-        });
-    }
-
-    function emitConnectPhaseCompleted(config: BlackBoxRallarConnectionConfig, phase: string, data?: unknown): void {
-        emitDiagnostic(config, 'rallar.browser.connect.phase_completed', {
-            phase,
-            ...toDiagnosticObject(data),
-        });
     }
 
     function requireState(): RuntimeState {
@@ -1682,9 +737,14 @@ function createBlackBoxRallarRuntimeInstallation(
     ): void {
         const key = authenticationKey(config, session.username);
         const sessionDiagnostic = toSessionDiagnostic(session);
+        const current = authenticationState;
+        const effectiveConfig =
+            current && current.key === key && isSameAuthenticationSession(current.session, sessionDiagnostic)
+                ? mergeBlackBoxRallarAuthenticationConfig(current.config, config)
+                : config;
         authenticationState = {
             key,
-            config,
+            config: effectiveConfig,
             session: sessionDiagnostic,
         };
         authenticationProvenance = {
@@ -1708,7 +768,7 @@ function createBlackBoxRallarRuntimeInstallation(
 
         authenticationState = {
             ...current,
-            config,
+            config: mergeBlackBoxRallarAuthenticationConfig(current.config, config),
         };
     }
 
@@ -1993,7 +1053,7 @@ function createBlackBoxRallarRuntimeInstallation(
     }
 
     function connectionKey(config: BlackBoxRallarConnectionConfig): string {
-        return JSON.stringify(blackBoxRallarConnectionTargetOf(config));
+        return blackBoxRallarConnectionOperationKeyOf(config);
     }
 
     function connectedTargetRejection(config: BlackBoxRallarConnectionConfig): Error | undefined {
@@ -2013,6 +1073,64 @@ function createBlackBoxRallarRuntimeInstallation(
         return decision.kind === 'reject'
             ? new Error(decision.reason)
             : undefined;
+    }
+
+    async function ensureCrdtLiveConnection(
+        config: BlackBoxRallarConnectionConfig,
+        transportStrategy: RallarCrdtTransportStrategy,
+    ): Promise<void> {
+        if (lifecycle.authenticationConfig()) {
+            await lifecycle.waitForAuthentication();
+            return await ensureCrdtLiveConnection(config, transportStrategy);
+        }
+        if (state) {
+            const decision = decideBlackBoxRallarLifecycleRequest(
+                {
+                    status: 'connected',
+                    activeTarget: blackBoxRallarConnectionTargetOf(state.config, state.session),
+                },
+                {
+                    kind: 'connect',
+                    target: blackBoxRallarConnectionTargetOf(config, state.session),
+                },
+            );
+            if (decision.kind === 'reject') {
+                throw new Error(decision.reason);
+            }
+        }
+        await lifecycle.runExclusive('crdt-live:' + connectionKey(config), async context => {
+            const queuedRejection = connectedTargetRejection(config);
+            if (queuedRejection) {
+                throw queuedRejection;
+            }
+            emitDiagnostic(config, 'rallar.browser.crdt.configure_started', {
+                transportStrategy,
+            });
+            const defaults = configureRallarConnection(config);
+            emitDiagnostic(config, 'rallar.browser.crdt.configure_completed', {
+                defaults,
+            });
+
+            if (!rallar.isConnected()) {
+                const session = await sessionForAuthentication(config);
+                context.assertCurrent();
+                await rallar.connect({
+                    timeoutMs: config.rallar.timeoutMs,
+                });
+                context.assertCurrent();
+                if (config.roomId) {
+                    await rallar.rooms.join(config.roomId, {
+                        timeoutMs: config.rallar.timeoutMs,
+                        scope: scopeOf(config),
+                    });
+                    context.assertCurrent();
+                }
+                emitDiagnostic(config, 'rallar.browser.crdt.connected', {
+                    session: toSessionDiagnostic(session),
+                    transportStrategy,
+                });
+            }
+        });
     }
 
     function connect(config: BlackBoxRallarConnectionConfig): Promise<BlackBoxRallarConnectDiagnostics> {
@@ -2037,1029 +1155,6 @@ function createBlackBoxRallarRuntimeInstallation(
             }
             return connectEffect(config, context);
         });
-    }
-
-    function normalizeSendInput(input: BlackBoxRallarSendInput | unknown): BlackBoxRallarSendInput {
-        if (
-            input &&
-            typeof input === 'object' &&
-            !Array.isArray(input) &&
-            ('data' in input ||
-                'laneId' in input ||
-                'roomId' in input ||
-                'roomRef' in input ||
-                'applicationId' in input ||
-                'workspaceId' in input ||
-                'scope' in input ||
-                'peerIds' in input ||
-                'nextHopPeerIds' in input ||
-                'remotePeerId' in input ||
-                'typeId' in input ||
-                'topicId' in input ||
-                'contextId' in input ||
-                'resourceId' in input)
-        ) {
-            return input as BlackBoxRallarSendInput;
-        }
-
-        return { data: input };
-    }
-
-    function normalizeMessagesRtcSendInput(input: BlackBoxRallarSendInput | unknown): BlackBoxRallarSendInput {
-        if (input && typeof input === 'object' && !Array.isArray(input)) {
-            return input as BlackBoxRallarSendInput;
-        }
-
-        return { payload: input };
-    }
-
-    function summarizeRealtimeSendResults(results: readonly RallarRealtimeSendResult[]): unknown {
-        const statuses: Record<string, number> = {};
-        const attentionResults = results.filter(entry => {
-            const status = entry.result.status;
-            statuses[status] = (statuses[status] ?? 0) + 1;
-            return status !== 'sent';
-        });
-
-        return {
-            total: results.length,
-            statuses,
-            peerIds: results.map(entry => entry.peerId),
-            attentionResults,
-        };
-    }
-
-    function realtimeSendStatusCount(summary: any, status: string): number {
-        return Number(summary?.statuses?.[status] ?? 0);
-    }
-
-    function emitRealtimeSendOutcomeDiagnostics(
-        config: BlackBoxRallarConnectionConfig,
-        diagnostics: BlackBoxRallarSendDiagnostics,
-    ): void {
-        const results = diagnostics.results ?? [];
-        const summary = summarizeRealtimeSendResults(results);
-
-        if (results.length === 0) {
-            emitDiagnostic(config, 'rallar.browser.realtime.peer_not_found', {
-                roomId: diagnostics.roomId,
-                laneId: diagnostics.laneId,
-                peerIds: diagnostics.peerIds,
-                health: diagnostics.health,
-                summary,
-            });
-        }
-
-        if (realtimeSendStatusCount(summary, 'closed') > 0) {
-            emitDiagnostic(config, 'rallar.browser.realtime.data_channel_not_open', {
-                roomId: diagnostics.roomId,
-                laneId: diagnostics.laneId,
-                peerIds: diagnostics.peerIds,
-                health: diagnostics.health,
-                summary,
-            });
-        }
-
-        if (
-            realtimeSendStatusCount(summary, 'queued') > 0 ||
-            realtimeSendStatusCount(summary, 'dropped') > 0 ||
-            realtimeSendStatusCount(summary, 'replaced') > 0 ||
-            realtimeSendStatusCount(summary, 'closed') > 0
-        ) {
-            emitDiagnostic(config, 'rallar.browser.realtime.send_result_attention', {
-                roomId: diagnostics.roomId,
-                laneId: diagnostics.laneId,
-                peerIds: diagnostics.peerIds,
-                health: diagnostics.health,
-                summary,
-            });
-        }
-    }
-
-    async function send(input: BlackBoxRallarSendInput | unknown): Promise<BlackBoxRallarSendDiagnostics> {
-        const runtimeState = requireState();
-        const { config } = runtimeState;
-        const lease = messagingController.lease();
-        messagingController.assertCurrent(lease, 'Rallar send completed after the runtime closed.');
-        try {
-            const transport = transportOf(config);
-            if (transport === 'messages.rtc') {
-                return await sendMessagesRtc(input, config, lease);
-            }
-
-            return await sendRealtime(input, config, lease);
-        } catch (error) {
-            emitError(config, `rallar.browser.${transportOf(config)}.send_failed`, error, {
-                transport: transportOf(config),
-            });
-            throw error;
-        }
-    }
-
-    async function sendRealtime(
-        input: BlackBoxRallarSendInput | unknown,
-        config: BlackBoxRallarConnectionConfig,
-        lease: BlackBoxRallarMessagingLease,
-    ): Promise<BlackBoxRallarSendDiagnostics> {
-        const transport = transportOf(config);
-        const normalized = normalizeSendInput(input);
-        const peerIds =
-            normalized.peerIds ??
-            (normalized.remotePeerId
-                ? [normalized.remotePeerId]
-                : config.remotePeerId
-                  ? [config.remotePeerId]
-                  : config.rallar.peerIds);
-        const laneId = normalized.laneId ?? laneIdOf(config);
-        const roomId = normalized.roomId ?? config.roomId;
-        const roomRef = roomRefOf(config, normalized);
-        const data = 'data' in normalized ? normalized.data : normalized.payload;
-
-        emitDiagnostic(config, 'rallar.browser.realtime.send_started', {
-            roomId,
-            roomRef,
-            ...scopeDiagnostics(config, normalized),
-            laneId,
-            peerIds,
-        });
-        const results = await rallar.realtime.sendJson({
-            data,
-            laneId,
-            roomId,
-            roomRef,
-            peerIds,
-            openTimeoutMs: normalized.openTimeoutMs ?? config.rallar.openTimeoutMs,
-            key: normalized.key,
-            maxAgeMs: normalized.maxAgeMs,
-        });
-        messagingController.assertCurrent(lease, 'Rallar send completed after the runtime closed.');
-        const diagnostics: BlackBoxRallarSendDiagnostics = {
-            status: results.length === 0 ? 'no-peers' : 'sent',
-            connection: config.connection,
-            actor: config.actor,
-            transport,
-            roomId,
-            ...scopeDiagnostics(config, normalized),
-            laneId,
-            peerIds,
-            results,
-            health: readHealth(config),
-        };
-        emitRealtimeSendOutcomeDiagnostics(config, diagnostics);
-        emitDiagnostic(config, 'rallar.browser.realtime.send_completed', diagnostics);
-        return diagnostics;
-    }
-
-    async function sendMessagesRtc(
-        input: BlackBoxRallarSendInput | unknown,
-        config: BlackBoxRallarConnectionConfig,
-        lease: BlackBoxRallarMessagingLease,
-    ): Promise<BlackBoxRallarSendDiagnostics> {
-        const transport = transportOf(config);
-        const normalized = normalizeMessagesRtcSendInput(input);
-        const typeId = normalized.typeId ?? typeIdOf(config);
-        const topicId = normalized.topicId ?? topicIdOf(config);
-        const roomId = normalized.roomId ?? config.roomId;
-        const roomRef = roomRefOf(config, normalized);
-        const contextId = normalized.contextId ?? config.rallar.contextId;
-        const resourceId = normalized.resourceId ?? config.rallar.resourceId;
-        const minSnapshotVersion = toOptionalNumber(normalized.minSnapshotVersion ?? config.rallar.minSnapshotVersion);
-        const nextHopPeerIds =
-            normalized.nextHopPeerIds ?? normalized.peerIds ?? config.rallar.nextHopPeerIds ?? config.rallar.peerIds;
-        const payload = 'payload' in normalized ? normalized.payload : normalized.data;
-
-        emitDiagnostic(config, 'rallar.browser.messages.rtc.send_started', {
-            roomId,
-            roomRef,
-            ...scopeDiagnostics(config, normalized),
-            nextHopPeerIds,
-            typeId,
-            topicId,
-            contextId,
-            resourceId,
-            minSnapshotVersion,
-        });
-        const messageInput: Record<string, unknown> = {
-            typeId,
-            topicId,
-            contextId,
-            resourceId,
-            roomId,
-            roomRef,
-            payload,
-            ttlHops: normalized.ttlHops ?? config.rallar.ttlHops,
-            ttlMs: normalized.ttlMs ?? config.rallar.ttlMs,
-            reliability: normalized.reliability ?? config.rallar.reliability,
-            ack: normalized.ack ?? config.rallar.ack,
-            ownership: normalized.ownership ?? config.rallar.ownership,
-            membershipEpoch: normalized.membershipEpoch ?? config.rallar.membershipEpoch,
-            minSnapshotVersion,
-            seq: normalized.seq ?? config.rallar.seq,
-            orderingKey: normalized.orderingKey ?? config.rallar.orderingKey,
-            nextHopPeerIds,
-            overlayId: normalized.overlayId ?? config.rallar.overlayId,
-            fanoutLimit: normalized.fanoutLimit ?? config.rallar.fanoutLimit,
-        };
-        const message = await rallar.messages.rtc.send(messageInput as any);
-        messagingController.assertCurrent(lease, 'Rallar send completed after the runtime closed.');
-
-        const diagnostics: BlackBoxRallarSendDiagnostics = {
-            status: 'sent',
-            connection: config.connection,
-            actor: config.actor,
-            transport,
-            roomId,
-            ...scopeDiagnostics(config, normalized),
-            nextHopPeerIds,
-            typeId,
-            topicId,
-            contextId,
-            resourceId,
-            minSnapshotVersion,
-            message,
-            health: readHealth(config),
-        };
-        emitDiagnostic(config, 'rallar.browser.messages.rtc.send_completed', diagnostics);
-        return diagnostics;
-    }
-
-    function normalizeWsSendInput(input: unknown): Record<string, unknown> {
-        return isRecord(input) ? input : { payload: input };
-    }
-
-    async function sendWs(input: unknown): Promise<BlackBoxRallarWsSendDiagnostics> {
-        const runtimeState = requireState();
-        const { config } = runtimeState;
-        const lease = messagingController.lease();
-        messagingController.assertCurrent(lease, 'Rallar send completed after the runtime closed.');
-        const normalized = normalizeWsSendInput(input);
-        const roomId = stringValue(normalized.roomId) ?? stringValue(normalized.groupId) ?? config.roomId;
-        const scope = wsScopeValue(normalized.scope) ?? (roomId ? 'room' : 'all');
-        const scopedInput = {
-            ...normalized,
-            ...(roomId ? { roomId } : {}),
-        } as BlackBoxRallarSendInput;
-        const roomRef = roomId ? roomRefOf(config, scopedInput) : undefined;
-        const typeId =
-            stringValue(normalized.typeId) ??
-            stringValue(normalized.topic) ??
-            stringValue(normalized.kind) ??
-            'rallar.black-box.ws.json';
-        const topicId = stringValue(normalized.topicId) ?? stringValue(normalized.topic) ?? typeId;
-        const contextId = stringValue(normalized.contextId) ?? roomId ?? scope;
-        const resourceId = stringValue(normalized.resourceId);
-        const minSnapshotVersion = toOptionalNumber(normalized.minSnapshotVersion ?? config.rallar.minSnapshotVersion);
-        const payload = 'payload' in normalized ? normalized.payload : 'data' in normalized ? normalized.data : input;
-        ensureWsMessageSubscription(config, typeId, topicId);
-
-        emit({
-            kind: 'diagnostic',
-            topic: 'rallar.browser.ws.send_started',
-            connection: config.connection,
-            actor: config.actor,
-            transport: 'ws',
-            roomId,
-            roomRef,
-            ...scopeDiagnostics(config, scopedInput),
-            typeId,
-            topicId,
-            contextId,
-            resourceId,
-            data: {
-                scope,
-                minSnapshotVersion,
-                wsStatus: wsStatusFor(),
-            },
-        });
-
-        try {
-            const result = await rallar.messages.ws.send({
-                typeId,
-                topicId,
-                contextId,
-                resourceId,
-                scope,
-                roomId,
-                roomRef,
-                payload,
-                minSnapshotVersion,
-                exceptPeerIds: maybeStringArray(normalized.exceptPeerIds),
-                ttlHops: toOptionalNumber(normalized.ttlHops ?? config.rallar.ttlHops),
-                ttlMs: toOptionalNumber(normalized.ttlMs ?? config.rallar.ttlMs),
-                reliability: normalized.reliability ?? config.rallar.reliability,
-                ack: normalized.ack ?? config.rallar.ack,
-                ownership: normalized.ownership ?? config.rallar.ownership,
-            } as any);
-            messagingController.assertCurrent(lease, 'Rallar send completed after the runtime closed.');
-            const diagnostics: BlackBoxRallarWsSendDiagnostics = {
-                status: 'sent',
-                connection: config.connection,
-                actor: config.actor,
-                transport: 'ws',
-                roomId,
-                roomRef,
-                ...scopeDiagnostics(config, scopedInput),
-                scope,
-                typeId,
-                topicId,
-                contextId,
-                resourceId,
-                minSnapshotVersion,
-                message: payload,
-                result,
-                wsStatus: wsStatusFor(),
-                rtcStatus: rtcStatusFor(config),
-            };
-            emit({
-                kind: 'diagnostic',
-                topic: 'rallar.browser.ws.send_completed',
-                connection: config.connection,
-                actor: config.actor,
-                transport: 'ws',
-                roomId,
-                roomRef,
-                ...scopeDiagnostics(config, scopedInput),
-                typeId,
-                topicId,
-                contextId,
-                resourceId,
-                data: diagnostics,
-            });
-            return diagnostics;
-        } catch (error) {
-            emitError(config, 'rallar.browser.ws.send_failed', error, {
-                transport: 'ws',
-                roomId,
-                roomRef,
-                typeId,
-                topicId,
-                contextId,
-                resourceId,
-                scope,
-            });
-            throw error;
-        }
-    }
-
-    function ensureWsMessageSubscription(
-        config: BlackBoxRallarConnectionConfig,
-        typeId: string,
-        topicId?: string,
-    ): void {
-        requireState();
-        const key = wsSelectorKey(typeId, topicId);
-        messagingController.ensureWsSubscription(key, () => {
-            const unsubscribe = rallar.messages.ws.onMessage(
-                {
-                    typeId,
-                    ...(topicId ? { topicId } : {}),
-                },
-                message => {
-                    emit({
-                        kind: 'message',
-                        topic: 'rallar.browser.ws.message',
-                        connection: config.connection,
-                        actor: config.actor,
-                        transport: 'ws',
-                        roomId: message.roomId ?? config.roomId,
-                        ...scopeDiagnostics(config),
-                        senderId: message.senderId,
-                        typeId: message.typeId,
-                        topicId: message.topicId,
-                        contextId: message.contextId,
-                        resourceId: message.resourceId,
-                        data: message.payload,
-                    });
-                },
-            );
-            emit({
-                kind: 'diagnostic',
-                topic: 'rallar.browser.ws.subscribed',
-                connection: config.connection,
-                actor: config.actor,
-                transport: 'ws',
-                roomId: config.roomId,
-                ...scopeDiagnostics(config),
-                typeId,
-                topicId,
-            });
-            return unsubscribe;
-        });
-    }
-
-    async function crdtOpen(
-        rawInput: BlackBoxRallarCrdtOpenInput | unknown,
-    ): Promise<BlackBoxRallarCrdtCommandDiagnostics> {
-        const input = normalizeCrdtOpenInput(rawInput);
-        const handle = crdtHandle(input);
-        const generation = lifecycle.generation();
-        let config: BlackBoxRallarConnectionConfig | undefined;
-        try {
-            if (!lifecycle.isCurrent(generation)) {
-                throw new Error('CRDT document open was cancelled because the Rallar runtime closed.');
-            }
-            const document = await crdtController.open(handle, async () => {
-                config = await ensureCrdtLiveConnection(input);
-                const opened = await rallar.crdt.open(input.name, toCrdtOpenOptions(input));
-                if (!lifecycle.isCurrent(generation)) {
-                    await opened.close();
-                    throw new Error('CRDT document open was cancelled because the Rallar runtime closed.');
-                }
-                return opened;
-            });
-            const diagnostics = toCrdtDiagnostics('opened', handle, document, {
-                transportStrategy: input.transport,
-            });
-            emitCrdtDiagnostic('rallar.browser.crdt.opened', handle, diagnostics, config);
-            return diagnostics;
-        } catch (error) {
-            emitError(config, 'rallar.browser.crdt.open_failed', error, {
-                handle,
-                transportStrategy: input.transport,
-            });
-            throw error;
-        }
-    }
-
-    async function runCrdtOperation<TResult>(
-        handle: string,
-        effect: (
-            document: RallarCrdtDocument<unknown, RallarCrdtOperationBatch>,
-            assertCurrent: () => void,
-        ) => Promise<TResult>,
-    ): Promise<TResult> {
-        const lease = crdtController.lease();
-        crdtController.assertCurrent(lease, 'CRDT operation completed after the runtime closed.');
-        return await crdtController.run(handle, async document => {
-            const assertCurrent = () =>
-                crdtController.assertCurrent(lease, 'CRDT operation completed after the runtime closed.');
-            const result = await effect(document, assertCurrent);
-            assertCurrent();
-            return result;
-        });
-    }
-
-    async function crdtApply(
-        input: BlackBoxRallarCrdtApplyInput | unknown,
-    ): Promise<BlackBoxRallarCrdtCommandDiagnostics> {
-        const record = asRecord(input);
-        const handle = crdtHandle(record);
-        const batch = record.batch as RallarCrdtOperationBatch | undefined;
-        if (!batch) {
-            throw new Error('crdt.apply requires batch.');
-        }
-
-        return await runCrdtOperation(handle, async (document, assertCurrent) => {
-            try {
-                const update = await document.applyLocal(batch);
-                assertCurrent();
-                const diagnostics = toCrdtDiagnostics('applied', handle, document, {
-                    update,
-                });
-                emitCrdtDiagnostic('rallar.browser.crdt.applied', handle, diagnostics);
-                return diagnostics;
-            } catch (error) {
-                emitError(undefined, 'rallar.browser.crdt.apply_failed', error, { handle });
-                throw error;
-            }
-        });
-    }
-
-    async function crdtRead(
-        input: BlackBoxRallarCrdtHandleInput | unknown,
-    ): Promise<BlackBoxRallarCrdtCommandDiagnostics> {
-        const handle = crdtHandle(input);
-        const document = requireCrdtDocument(handle);
-        const diagnostics = toCrdtDiagnostics('read', handle, document, {
-            value: document.read(),
-        });
-        emitCrdtDiagnostic('rallar.browser.crdt.read', handle, diagnostics);
-        return diagnostics;
-    }
-
-    async function crdtSync(
-        input: BlackBoxRallarCrdtSyncInput | unknown,
-    ): Promise<BlackBoxRallarCrdtCommandDiagnostics> {
-        const record = asRecord(input);
-        const handle = crdtHandle(record);
-        const options: RallarCrdtSyncOptions = {
-            ...(stringValue(record.reason) ? { reason: stringValue(record.reason) } : {}),
-            ...(toCrdtTransport(record.transport) ? { transport: toCrdtTransport(record.transport) } : {}),
-        };
-
-        return await runCrdtOperation(handle, async (document, assertCurrent) => {
-            try {
-                const result = await document.sync(options);
-                assertCurrent();
-                const diagnostics = toCrdtDiagnostics('synced', handle, document, {
-                    result,
-                    transportStrategy: toCrdtTransport(record.transport),
-                });
-                emitCrdtDiagnostic('rallar.browser.crdt.synced', handle, diagnostics);
-                return diagnostics;
-            } catch (error) {
-                emitError(undefined, 'rallar.browser.crdt.sync_failed', error, { handle });
-                throw error;
-            }
-        });
-    }
-
-    async function crdtHealth(
-        input: BlackBoxRallarCrdtHandleInput | unknown,
-    ): Promise<BlackBoxRallarCrdtCommandDiagnostics> {
-        const handle = crdtHandle(input);
-        const document = requireCrdtDocument(handle);
-        const diagnostics = toCrdtDiagnostics('health', handle, document);
-        emitCrdtDiagnostic('rallar.browser.crdt.health', handle, diagnostics);
-        return diagnostics;
-    }
-
-    async function crdtWaitEffect(
-        input: BlackBoxRallarCrdtWaitInput,
-        document: RallarCrdtDocument<unknown, RallarCrdtOperationBatch>,
-        assertCurrent: () => void,
-    ): Promise<BlackBoxRallarCrdtCommandDiagnostics> {
-        const timeoutMs = input.timeoutMs ?? 10_000;
-        const intervalMs = input.intervalMs ?? 250;
-        const stableForMs = input.stableForMs ?? 0;
-        const startEpochMs = now();
-        const deadlineEpochMs = startEpochMs + timeoutMs;
-        const syncOptions: RallarCrdtSyncOptions | undefined =
-            input.sync && typeof input.sync === 'object'
-                ? {
-                      ...(input.sync.reason ? { reason: input.sync.reason } : {}),
-                      ...(input.sync.transport ? { transport: input.sync.transport } : {}),
-                  }
-                : undefined;
-        let attempts = 0;
-        let stableSinceEpochMs: number | undefined;
-        let lastSyncResult: unknown;
-        let lastValue: unknown;
-        let lastHealth: unknown;
-
-        emitCrdtDiagnostic('rallar.browser.crdt.waiting', input.handle, {
-            status: 'waiting',
-            handle: input.handle,
-            ref: document.ref,
-            timeoutMs,
-            intervalMs,
-            stableForMs,
-            conditions: input.conditions,
-            sync: input.sync,
-        });
-
-        try {
-            while (true) {
-                attempts += 1;
-                if (syncOptions) {
-                    lastSyncResult = await document.sync(syncOptions);
-                    assertCurrent();
-                }
-
-                lastValue = document.read();
-                lastHealth = document.health();
-                const currentEpochMs = now();
-                const matched = input.conditions.every(condition =>
-                    crdtWaitConditionMatches(condition, lastValue, lastHealth),
-                );
-
-                if (matched) {
-                    stableSinceEpochMs ??= currentEpochMs;
-                    if (stableForMs <= 0 || currentEpochMs - stableSinceEpochMs >= stableForMs) {
-                        assertCurrent();
-                        const diagnostics = toCrdtDiagnostics('wait_matched', input.handle, document, {
-                            value: lastValue,
-                            result: {
-                                matched: true,
-                                matchedAtEpochMs: currentEpochMs,
-                            },
-                            attempts,
-                            waitedMs: currentEpochMs - startEpochMs,
-                            stableForMs,
-                            conditions: input.conditions,
-                            lastSyncResult,
-                        });
-                        emitCrdtDiagnostic('rallar.browser.crdt.wait_matched', input.handle, diagnostics);
-                        return diagnostics;
-                    }
-                } else {
-                    stableSinceEpochMs = undefined;
-                }
-
-                if (currentEpochMs >= deadlineEpochMs) {
-                    throw new Error('Timed out waiting for CRDT conditions on handle: ' + input.handle);
-                }
-
-                await wait(Math.min(intervalMs, Math.max(0, deadlineEpochMs - currentEpochMs)));
-                assertCurrent();
-            }
-        } catch (error) {
-            emitError(undefined, 'rallar.browser.crdt.wait_failed', error, {
-                handle: input.handle,
-                attempts,
-                waitedMs: now() - startEpochMs,
-                stableForMs,
-                conditions: input.conditions,
-                lastValue,
-                lastHealth,
-                lastSyncResult,
-            });
-            throw error;
-        }
-    }
-
-    async function crdtWait(
-        rawInput: BlackBoxRallarCrdtWaitInput | unknown,
-    ): Promise<BlackBoxRallarCrdtCommandDiagnostics> {
-        const input = normalizeCrdtWaitInput(rawInput);
-        return await runCrdtOperation(input.handle, (document, assertCurrent) =>
-            crdtWaitEffect(input, document, assertCurrent),
-        );
-    }
-
-    async function crdtUndo(
-        input: BlackBoxRallarCrdtUndoRedoInput | unknown,
-    ): Promise<BlackBoxRallarCrdtCommandDiagnostics> {
-        const record = asRecord(input);
-        const handle = crdtHandle(record);
-        const targetOperationGroupId = stringValue(record.targetOperationGroupId);
-        const operations = Array.isArray(record.operations)
-            ? (record.operations as readonly RallarCrdtOperation[])
-            : undefined;
-        if (!targetOperationGroupId || !operations) {
-            throw new Error('crdt.undo requires targetOperationGroupId and operations.');
-        }
-
-        return await runCrdtOperation(handle, async (document, assertCurrent) => {
-            try {
-                const update = await document.undoOperationGroup({
-                    targetOperationGroupId,
-                    operations,
-                    ...(stringValue(record.operationGroupId)
-                        ? {
-                              operationGroupId: stringValue(record.operationGroupId),
-                          }
-                        : {}),
-                });
-                assertCurrent();
-                const diagnostics = toCrdtDiagnostics('undone', handle, document, {
-                    update,
-                });
-                emitCrdtDiagnostic('rallar.browser.crdt.undone', handle, diagnostics);
-                return diagnostics;
-            } catch (error) {
-                emitError(undefined, 'rallar.browser.crdt.undo_failed', error, { handle });
-                throw error;
-            }
-        });
-    }
-
-    async function crdtRedo(
-        input: BlackBoxRallarCrdtUndoRedoInput | unknown,
-    ): Promise<BlackBoxRallarCrdtCommandDiagnostics> {
-        const record = asRecord(input);
-        const handle = crdtHandle(record);
-        const targetOperationGroupId = stringValue(record.targetOperationGroupId);
-        const operations = Array.isArray(record.operations)
-            ? (record.operations as readonly RallarCrdtOperation[])
-            : undefined;
-        if (!targetOperationGroupId || !operations) {
-            throw new Error('crdt.redo requires targetOperationGroupId and operations.');
-        }
-
-        return await runCrdtOperation(handle, async (document, assertCurrent) => {
-            try {
-                const update = await document.redoOperationGroup({
-                    targetOperationGroupId,
-                    operations,
-                    ...(stringValue(record.operationGroupId)
-                        ? {
-                              operationGroupId: stringValue(record.operationGroupId),
-                          }
-                        : {}),
-                });
-                assertCurrent();
-                const diagnostics = toCrdtDiagnostics('redone', handle, document, {
-                    update,
-                });
-                emitCrdtDiagnostic('rallar.browser.crdt.redone', handle, diagnostics);
-                return diagnostics;
-            } catch (error) {
-                emitError(undefined, 'rallar.browser.crdt.redo_failed', error, { handle });
-                throw error;
-            }
-        });
-    }
-
-    async function crdtClose(
-        input: BlackBoxRallarCrdtHandleInput | unknown,
-    ): Promise<BlackBoxRallarCrdtCommandDiagnostics> {
-        const handle = crdtHandle(input);
-        return await runCrdtOperation(handle, async (document, assertCurrent) => {
-            const diagnostics = toCrdtDiagnostics('closed', handle, document);
-            await document.close();
-            assertCurrent();
-            crdtController.delete(handle);
-            emitCrdtDiagnostic('rallar.browser.crdt.closed', handle, diagnostics);
-            return diagnostics;
-        });
-    }
-
-    async function crdtDestroy(
-        input: BlackBoxRallarCrdtHandleInput | unknown,
-    ): Promise<BlackBoxRallarCrdtCommandDiagnostics> {
-        const handle = crdtHandle(input);
-        return await runCrdtOperation(handle, async (document, assertCurrent) => {
-            const diagnostics = toCrdtDiagnostics('destroyed', handle, document);
-            await document.destroy();
-            assertCurrent();
-            crdtController.delete(handle);
-            emitCrdtDiagnostic('rallar.browser.crdt.destroyed', handle, diagnostics);
-            return diagnostics;
-        });
-    }
-
-    async function directorAppoint(
-        input: BlackBoxRallarDirectorAppointInput | unknown,
-    ): Promise<BlackBoxRallarDirectorCommandDiagnostics> {
-        const runtimeState = requireState();
-        const config = runtimeState.config;
-        const lease = directorController.lease();
-        assertDirectorCurrent(lease);
-        const normalized = normalizeDirectorAppointInput(input);
-        const target = toDirectorTarget(normalized, config);
-        const directorStatus = await rallar.director.appoint(
-            target as any,
-            {
-                heartbeatTtlMs: normalized.heartbeatTtlMs,
-                scope: toDirectorScope(normalized, config),
-                timeoutMs: normalized.timeoutMs,
-            } as any,
-        );
-        assertDirectorCurrent(lease);
-        const diagnostics = directorStatusDiagnostics('appointed', normalized, directorStatus, config);
-        emitDirectorDiagnostic('rallar.browser.director.appointed', undefined, diagnostics, config);
-        return diagnostics;
-    }
-
-    async function directorResign(
-        input: BlackBoxRallarDirectorRoomInput | unknown,
-    ): Promise<BlackBoxRallarDirectorCommandDiagnostics> {
-        const runtimeState = requireState();
-        const config = runtimeState.config;
-        const lease = directorController.lease();
-        assertDirectorCurrent(lease);
-        const normalized = normalizeDirectorRoomInput(input);
-        const target = toDirectorTarget(normalized, config);
-        const directorStatus = await rallar.director.resign(
-            target as any,
-            {
-                scope: toDirectorScope(normalized, config),
-                timeoutMs: normalized.timeoutMs,
-            } as any,
-        );
-        assertDirectorCurrent(lease);
-        const diagnostics = directorStatusDiagnostics('resigned', normalized, directorStatus, config);
-        emitDirectorDiagnostic('rallar.browser.director.resigned', undefined, diagnostics, config);
-        return diagnostics;
-    }
-
-    async function directorStatus(
-        input: BlackBoxRallarDirectorStatusInput | unknown,
-    ): Promise<BlackBoxRallarDirectorCommandDiagnostics> {
-        const runtimeState = requireState();
-        const config = runtimeState.config;
-        const lease = directorController.lease();
-        assertDirectorCurrent(lease);
-        const normalized = normalizeDirectorStatusInput(input);
-        const target = toDirectorTarget(normalized, config);
-        if (normalized.refresh) {
-            await rallar.rooms.refresh({
-                scope: toDirectorScope(normalized, config),
-                timeoutMs: normalized.timeoutMs,
-            } as any);
-        }
-        assertDirectorCurrent(lease);
-        const status = rallar.director.status(target as any, {
-            now: normalized.now,
-        });
-        const diagnostics = directorStatusDiagnostics('status', normalized, status, config);
-        emitDirectorDiagnostic('rallar.browser.director.status', undefined, diagnostics, config);
-        return diagnostics;
-    }
-
-    async function directorRelayStart(
-        input: BlackBoxRallarDirectorRelayStartInput | unknown,
-    ): Promise<BlackBoxRallarDirectorCommandDiagnostics> {
-        const runtimeState = requireState();
-        const config = runtimeState.config;
-        const lease = directorController.lease();
-        assertDirectorCurrent(lease);
-        const normalized = normalizeDirectorRelayStartInput(input);
-        if (directorController.handles().includes(normalized.handle)) {
-            throw new Error('Director relay handle is already open: ' + normalized.handle);
-        }
-
-        let entry!: DirectorRelayState;
-        const relay = rallar.director.createRelay<unknown, BlackBoxRallarDirectorOutputRecord, unknown>({
-            roomId: normalized.roomId ?? config.roomId,
-            roomRef: toDirectorRoomRef(normalized, config) as any,
-            laneId: normalized.laneId,
-            topicId: normalized.topicId,
-            intentTypeId: normalized.intentTypeId,
-            outputTypeId: normalized.outputTypeId,
-            heartbeatTypeId: normalized.heartbeatTypeId,
-            snapshotTypeId: normalized.snapshotTypeId,
-            syncRequestTypeId: normalized.syncRequestTypeId,
-            heartbeatIntervalMs: normalized.heartbeatIntervalMs,
-            snapshotIntervalMs: normalized.snapshotIntervalMs,
-            readSnapshot: () => directorRelaySnapshot(entry),
-            onIntent: (message: RallarDirectorRelayMessage<unknown>) => {
-                assertDirectorCurrent(lease);
-                const nextSequence = entry.sequence + 1;
-                const output: BlackBoxRallarDirectorOutputRecord = {
-                    kind: 'black-box-director-output',
-                    intentId: intentIdFromPayload(message.data, `intent-${nextSequence}`),
-                    sequence: nextSequence,
-                    senderId: message.senderId,
-                    directorSessionId: entry.relay.status().appointment?.sessionId,
-                    directorPrincipalId: entry.relay.status().appointment?.principalId,
-                    epoch: message.envelope.epoch,
-                    receivedAtEpochMs: message.receivedAtEpochMs,
-                    payload: message.data,
-                };
-                entry.sequence = nextSequence;
-                entry.acceptedIntents.push({
-                    intentId: output.intentId,
-                    senderId: message.senderId,
-                    epoch: message.envelope.epoch,
-                    receivedAtEpochMs: message.receivedAtEpochMs,
-                    payload: message.data,
-                });
-                entry.outputs.push(output);
-                emitDirectorDiagnostic(
-                    'rallar.browser.director.intent_received',
-                    entry.handle,
-                    {
-                        intent: entry.acceptedIntents.at(-1),
-                        output,
-                    },
-                    config,
-                );
-                return output;
-            },
-            onOutput: (message: RallarDirectorRelayMessage<BlackBoxRallarDirectorOutputRecord>) => {
-                assertDirectorCurrent(lease);
-                entry.outputs.push(message.data);
-                emitDirectorDiagnostic(
-                    'rallar.browser.director.output_received',
-                    entry.handle,
-                    {
-                        output: message.data,
-                        senderId: message.senderId,
-                        epoch: message.envelope.epoch,
-                        receivedAtEpochMs: message.receivedAtEpochMs,
-                    },
-                    config,
-                );
-            },
-            onSnapshot: (message: RallarDirectorRelayMessage<unknown>) => {
-                assertDirectorCurrent(lease);
-                entry.snapshots.push(message.data);
-                emitDirectorDiagnostic(
-                    'rallar.browser.director.snapshot_received',
-                    entry.handle,
-                    {
-                        snapshot: message.data,
-                        senderId: message.senderId,
-                        epoch: message.envelope.epoch,
-                        receivedAtEpochMs: message.receivedAtEpochMs,
-                    },
-                    config,
-                );
-            },
-            onSyncRequest: (message: RallarDirectorRelayMessage<unknown>) => {
-                assertDirectorCurrent(lease);
-                entry.syncRequests.push({
-                    senderId: message.senderId,
-                    epoch: message.envelope.epoch,
-                    receivedAtEpochMs: message.receivedAtEpochMs,
-                    payload: message.data,
-                });
-                emitDirectorDiagnostic(
-                    'rallar.browser.director.sync_request_received',
-                    entry.handle,
-                    {
-                        syncRequest: entry.syncRequests.at(-1),
-                    },
-                    config,
-                );
-            },
-        });
-
-        entry = {
-            handle: normalized.handle,
-            input: normalized,
-            relay,
-            acceptedIntents: [],
-            outputs: [],
-            snapshots: [],
-            syncRequests: [],
-            sequence: 0,
-        };
-        directorController.add(normalized.handle, entry);
-        const status = relay.status();
-        const diagnostics = directorStatusDiagnostics('relay_started', normalized, status, config, {
-            handle: normalized.handle,
-            relay: {
-                handle: normalized.handle,
-                topicId: normalized.topicId,
-                intentTypeId: normalized.intentTypeId,
-                outputTypeId: normalized.outputTypeId,
-                heartbeatTypeId: normalized.heartbeatTypeId,
-                snapshotTypeId: normalized.snapshotTypeId,
-                syncRequestTypeId: normalized.syncRequestTypeId,
-            },
-        });
-        emitDirectorDiagnostic('rallar.browser.director.relay_started', normalized.handle, diagnostics, config);
-        return diagnostics;
-    }
-
-    async function directorIntent(
-        input: BlackBoxRallarDirectorIntentInput | unknown,
-    ): Promise<BlackBoxRallarDirectorCommandDiagnostics> {
-        const runtimeState = requireState();
-        const config = runtimeState.config;
-        const lease = directorController.lease();
-        assertDirectorCurrent(lease);
-        const normalized = normalizeDirectorIntentInput(input);
-        const relay = requireDirectorRelay(normalized.handle);
-        const sendResult = await relay.relay.sendIntent(normalized.intent);
-        assertDirectorCurrent(lease);
-        const diagnostics = directorStatusDiagnostics('intent_sent', relay.input, relay.relay.status(), config, {
-            handle: normalized.handle,
-            sendResult,
-        });
-        emitDirectorDiagnostic(
-            'rallar.browser.director.intent_sent',
-            normalized.handle,
-            {
-                ...diagnostics,
-                intent: normalized.intent,
-            },
-            config,
-        );
-        return diagnostics;
-    }
-
-    async function directorSyncRequest(
-        input: BlackBoxRallarDirectorSyncRequestInput | unknown,
-    ): Promise<BlackBoxRallarDirectorCommandDiagnostics> {
-        const runtimeState = requireState();
-        const config = runtimeState.config;
-        const lease = directorController.lease();
-        assertDirectorCurrent(lease);
-        const normalized = normalizeDirectorSyncRequestInput(input);
-        const relay = requireDirectorRelay(normalized.handle);
-        const sendResult = await relay.relay.requestSync(normalized.payload);
-        assertDirectorCurrent(lease);
-        const diagnostics = directorStatusDiagnostics('sync_requested', relay.input, relay.relay.status(), config, {
-            handle: normalized.handle,
-            sendResult,
-        });
-        emitDirectorDiagnostic(
-            'rallar.browser.director.sync_requested',
-            normalized.handle,
-            {
-                ...diagnostics,
-                payload: normalized.payload,
-            },
-            config,
-        );
-        return diagnostics;
-    }
-
-    async function directorRelayStop(
-        input: BlackBoxRallarDirectorHandleInput | unknown,
-    ): Promise<BlackBoxRallarDirectorCommandDiagnostics> {
-        const runtimeState = requireState();
-        const config = runtimeState.config;
-        const lease = directorController.lease();
-        assertDirectorCurrent(lease);
-        const normalized = normalizeDirectorHandleInput(input);
-        const relay = requireDirectorRelay(normalized.handle);
-        const status = relay.relay.status();
-        relay.relay.stop();
-        directorController.delete(normalized.handle);
-        const diagnostics = directorStatusDiagnostics('relay_stopped', relay.input, status, config, {
-            handle: normalized.handle,
-            acceptedIntentCount: relay.acceptedIntents.length,
-            outputCount: relay.outputs.length,
-            snapshotCount: relay.snapshots.length,
-            syncRequestCount: relay.syncRequests.length,
-        });
-        emitDirectorDiagnostic('rallar.browser.director.relay_stopped', normalized.handle, diagnostics, config);
-        return diagnostics;
     }
 
     async function closeEffect(preparation: RuntimeClosePreparation): Promise<BlackBoxRallarCloseDiagnostics> {
@@ -3087,57 +1182,12 @@ function createBlackBoxRallarRuntimeInstallation(
                 emitError(config, 'rallar.browser.cleanup.unsubscribe_failed', error);
             }
 
-            for (const [handle, relay] of directorController.entries()) {
-                try {
-                    relay.relay.stop();
-                    directorController.delete(handle);
-                    if (config) {
-                        emitDirectorDiagnostic(
-                            'rallar.browser.director.relay_stopped',
-                            handle,
-                            {
-                                status: 'relay_stopped',
-                                handle,
-                                reason: 'runtime-close',
-                                acceptedIntentCount: relay.acceptedIntents.length,
-                                outputCount: relay.outputs.length,
-                                snapshotCount: relay.snapshots.length,
-                                syncRequestCount: relay.syncRequests.length,
-                            },
-                            config,
-                        );
-                    }
-                } catch (error) {
-                    cleanupErrors.push(serializeError(error));
-                    emitError(config, 'rallar.browser.director.relay_stop_failed', error, {
-                        handle,
-                        reason: 'runtime-close',
-                    });
-                }
+            for (const error of directorController.closeAll(config)) {
+                cleanupErrors.push(serializeError(error));
             }
 
-            for (const [handle, document] of crdtController.entries()) {
-                try {
-                    await document.close();
-                    crdtController.delete(handle);
-                    emitCrdtDiagnostic(
-                        'rallar.browser.crdt.closed',
-                        handle,
-                        {
-                            status: 'closed',
-                            handle,
-                            ref: document.ref,
-                            reason: 'runtime-close',
-                        },
-                        config,
-                    );
-                } catch (error) {
-                    cleanupErrors.push(serializeError(error));
-                    emitError(config, 'rallar.browser.crdt.close_failed', error, {
-                        handle,
-                        reason: 'runtime-close',
-                    });
-                }
+            for (const error of await crdtController.closeAll(config)) {
+                cleanupErrors.push(serializeError(error));
             }
 
             if (config?.roomId && config.rallar.leaveRoomOnClose !== false) {
@@ -3291,37 +1341,18 @@ function createBlackBoxRallarRuntimeInstallation(
             health: config ? readHealth(config) : [],
             ...(rtcDiagnostics !== undefined ? { rtcDiagnostics } : {}),
             ...(rtcDiagnosticsError !== undefined ? { rtcDiagnosticsError } : {}),
-            crdt: crdtRuntimeSummary(),
-            director: directorRelaySummary(),
+            crdt: crdtController.summary(),
+            director: directorController.summary(),
         };
     }
 
     const runtime: BlackBoxRallarRuntime = {
         authenticate,
         connect,
-        send,
-        sendWs,
-        crdt: {
-            open: crdtOpen,
-            apply: crdtApply,
-            read: crdtRead,
-            sync: crdtSync,
-            health: crdtHealth,
-            wait: crdtWait,
-            undo: crdtUndo,
-            redo: crdtRedo,
-            close: crdtClose,
-            destroy: crdtDestroy,
-        },
-        director: {
-            appoint: directorAppoint,
-            resign: directorResign,
-            status: directorStatus,
-            relayStart: directorRelayStart,
-            intent: directorIntent,
-            syncRequest: directorSyncRequest,
-            relayStop: directorRelayStop,
-        },
+        send: messagingController.send,
+        sendWs: messagingController.sendWs,
+        crdt: crdtController,
+        director: directorController,
         close,
         health,
     };
